requirements.txt
wikipedia
openai
langchain
streamlit
python-dotenv
numexpr